test: split omnibus test modules
